refactor(Apple) : 애플 관련 서비스 패키지 분리
