--- src/main/java/org/noostak/auth/application/AppleService.java
@@ -1,5 +1,0 @@
-package org.noostak.auth.application;
-
-public interface AppleService extends OauthService {
-
-}
