+++ src/main/java/org/noostak/auth/application/apple/AppleApi.java
@@ -1,4 +1,4 @@
-package org.noostak.auth.application;
+package org.noostak.auth.application.apple;
 
 
 import lombok.AllArgsConstructor;
