+++ src/main/java/org/noostak/auth/application/apple/AppleService.java
@@ -1,0 +1,7 @@
+package org.noostak.auth.application.apple;
+
+import org.noostak.auth.application.OauthService;
+
+public interface AppleService extends OauthService {
+
+}
